Refactors TraceReader classes to reduce the number of abstract classes.
Simplifies Google Cluster Data constructors.
